Temporary fix of the bugged fetch magnet mechanism
This commit is a workaround for problems with the incorrect behavior of some trackers and the inability to change the priority of files (issues #93, #111) that are contained in the jlibtorrent library because the developers at the moment don't intend to fix them in the near future. The solution is simple: remove TorrentHandle of the magnet after receiving the metadata and create it again. A similar method is used in the synchronous method SessionManager.fetchMagnet