Migrate from Tray to shared preferences
In version 1.0, this app was multiprocess, so the Tray library was used as the preferences store. For a long time the app is oneprocess, so Tray is a bad legacy, which required deletion from the source code. Now settings are stored in shared preferences, settings from the Tray are automatically imported by the first time app is running.