More improve large torrents support
The main problem is that the Binder transaction buffer has a limited fixed size, currently 1Mb, and data with a large size that are transmitted through intent or which are stored in fragments are throws TransactionTooLargeException, see https://developer.android.com/reference/android/os/TransactionTooLargeException.

As a rule, the large data size is caused by a large number of files in the torrent (`AdvanceStateParcel`, `Torrent`, `BencodeFileTree` objects and etc).

For this we have taken the following steps:
 1) Using greenrobot EventBus for exchange messages between the service and fragments.
 2) To save large objects in fragments, uses a special 'retained' fragment, see https://github.com/codepath/android_guides/wiki/Handling-Configuration-Changes#retaining-fragments.
 3) And also EventBus sticky event for `AddTorrentActivity`. This allows you to add torrents with a large number of files.

In general, all steps are aimed at avoiding intents and saving state in fragments for large data.