Optimize file tree model for big torrents; fix race condition for file tree
[1]

Creating a file tree (in detail dialog) for a torrent with a large number of
files (for example, 10000) and a complex dir structure took a very long time,
it could reach 10 minutes or more.
To reduce time, the following optimizations were performed:
 - Refusal to depth-first search for tree leaves;
   now access to the leaves takes O(1).
 - Priority change and selections operations are combined,
   so as not to repeat similar actions twice.
 - By default, a parent updates its priority and selection only when
   all children are updated. This allows to reduce the number of
   children  enumeration to obtain their state, so that the parent is
   able to update its state.

These optimizations have reduced tree creation time to a few milliseconds.

[2]

In the detail dialog, race condition was fixed, that appeared while updating
metainfo, at the time the file tree was created.